+++ src/main/java/com/restaurante/proyecto/Proyectosoftware2Application.java
@@ -1,16 +1,12 @@
 package com.restaurante.proyecto;
 
+import com.restaurante.proyecto.models.dao.GreetingRepository;
+import com.restaurante.proyecto.models.entity.Greeting;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import javax.persistence.*;
-import java.util.List;
 
 @SpringBootApplication
 public class Proyectosoftware2Application {
@@ -20,67 +16,11 @@ public class Proyectosoftware2Application {
     }
 
     @Bean
-    ApplicationRunner applicationRunner(Greeting.GreetingRepository greetingRepository) {
+    ApplicationRunner applicationRunner(GreetingRepository greetingRepository) {
         return args -> {
             greetingRepository.save(new Greeting("Hola"));
             greetingRepository.save(new Greeting("Hi"));
         };
 
-    }
-}
-
-@RestController
-class HelloController {
-    private final Greeting.GreetingRepository greetingRepository;
-
-    HelloController(Greeting.GreetingRepository greetingRepository) {
-        this.greetingRepository = greetingRepository;
-    }
-
-    @GetMapping("/")
-    String hello() {
-        return "Hello World";
-    }
-
-    @GetMapping("/greetings")
-    List<Greeting> greetings() {
-        return (List<Greeting>) greetingRepository.findAll();
-    }
-
-}
-
-@Entity
-class Greeting {
-    @Id
-    @GeneratedValue
-    private Long id;
-    @Column
-    private String message;
-
-    public Greeting() {
-    }
-
-    public Greeting(String message) {
-        this.message = message;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    @Component
-    interface GreetingRepository extends CrudRepository<Greeting, Long> {
     }
 }
