+++ src/main/java/com/restaurante/proyecto/models/dao/GreetingRepository.java
@@ -1,0 +1,8 @@
+package com.restaurante.proyecto.models.dao;
+
+
+import com.restaurante.proyecto.models.entity.Greeting;
+import org.springframework.data.repository.CrudRepository;
+
+public interface GreetingRepository extends CrudRepository<Greeting, Long> {
+}
