+++ src/main/java/com/restaurante/proyecto/models/entity/Greeting.java
@@ -1,0 +1,40 @@
+package com.restaurante.proyecto.models.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public
+class Greeting {
+    @Id
+    @GeneratedValue
+    private Long id;
+    @Column
+    private String message;
+
+    public Greeting() {
+    }
+
+    public Greeting(String message) {
+        this.message = message;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
+
