+++ src/main/resources/application.properties
@@ -1,2 +1,10 @@
-
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.datasource.url=jdbc:postgresql://ec2-34-197-171-33.compute-1.amazonaws.com:5432/d9te6vdvmh3khi
+spring.datasource.username=hkpnjyrfoiqmqp
+spring.datasource.password=d27b1340724e50e2ff131ec65da2c58a62e58a729a2d0cc102a062fcce6ce12a
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.jpa.generate-ddl=true
+spring.jpa.show-sql=false
+spring.jpa.hibernate.ddl-auto=create
+spring.data.rest.basePath=/api
+spring.devtools.restart.log-condition-evaluation-delta=false
